feat: 860: organized files with their respective css files into folders
