+++ src/federation-info-plugin/FederationInfo/FederationInfo.module.scss
