+++ src/federation-info-plugin/views/SchemaView/SchemaView.jsx
