+++ src/federation-info-plugin/views/SchemaView/SchemaView.module.scss
